migrating redigo for connection pool
