+++ worker.go
@@ -1,0 +1,11 @@
+package gocelery
+
+// Worker represents distributed task worker
+type Worker struct {
+	broker CeleryBroker
+}
+
+// Run - start workers
+func (w *Worker) Run() error {
+	return nil
+}
